# docs/OpenAPIv3_swagger.yaml
components:
  schemas:
    domain.APIKey:
      properties:
        created_at:
          type: string
        id:
          type: integer
        label:
          type: string
        organization_id:
          type: integer
        revoked_at:
          type: string
      required:
      - created_at
      - id
      - label
      - organization_id
      - revoked_at
      type: object
...
    handler.GenerateKeyRequest:
      properties:
        label:
          type: string
      required:
      - label
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: x-organization-api-key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: chatlogger-api-go+swagger@kjanat.com
    name: ChatLogger
    url: https://github.com/kjanat/chatlogger-api-go/issues
  description: API for logging and managing chat sessions.
  license:
    name: MIT License
    url: https://github.com/kjanat/chatlogger-api-go/blob/master/LICENSE
  termsOfService: https://github.com/kjanat/chatlogger-api-go#terms-of-service
  title: ChatLogger API (Go)
  version: ""
openapi: 3.1.0
paths:
...
  /v1/orgs/me/apikeys:
    get:
      description: Lists all API keys for the organization associated with the authenticated
        user.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.APIKey'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list API keys
      security:
      - BearerAuth: []
      summary: List API Keys
      tags:
      - API Keys (Admin)
    post:
      description: Generates a new API key for the organization associated with the
        authenticated user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.GenerateKeyRequest'
        description: API Key Label
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key generated successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to generate API key
      security:
      - BearerAuth: []
      summary: Generate API Key
      tags:
      - API Keys (Admin)
  /v1/orgs/me/apikeys/{id}:
    delete:
      description: Revokes an existing API key by its ID.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key revoked successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid API key ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to revoke API key
      security:
      - BearerAuth: []
      summary: Revoke API Key
      tags:
      - API Keys (Admin)
  /v1/orgs/me/apikeys/{id}/delete:
    delete:
      description: Deletes an API key permanently. Use RevokeKey for standard deactivation.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key deleted successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid API key ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to delete API key
      security:
      - BearerAuth: []
      summary: Delete API Key (Not typically exposed, Revoke is preferred)
      tags:
      - API Keys (Admin)
...